Currently, if you have a negative Duration, the String returned from _getTimeString appears as -1:-2:-47.
This change checks if the Duration value passed in is negative.
If it is, we set the timeIsNegative bool to true and convert the
Duration value to a positive number using .abs(). If the Duration
is negative then we return a String with a prefixed with '-'.

With this change, the String will appears as -1:02:47.

ps-id: e7372e7d-3ff1-428a-96f3-879b40ee32a6